sandbox/jmillan/land_cover_classification_unet/requirements.txt
# Core scientific packages
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.4.0

# Deep learning frameworks - Mac Silicon compatible versions
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
torch>=2.1.0
torchvision>=0.16.0

# Utilities
tensorboard>=2.15.0
tqdm>=4.65.0
future>=0.18.3

# Optional dependencies
scikit-build>=0.17.0
